==> src/2_Features_Extraction.ipynb <==
import os
import pandas as pd
import numpy as np
from utilities.statsAPI import StatsAPI, FeatureGenerator, create_training_dataset
pd.set_option('display.max_columns', None)

## Create Training Dataset Using StatsAPI

df = pd.read_csv("../data/processed/Preprocessed_Data.csv")
df['match_date'] = pd.to_datetime(df['match_date'])

training_df = create_training_dataset(df)

## Dataset Lookthrough - Jannik Sinner Last 10 Matches 

sinner = training_df[ ((training_df["P1_ID"] == 206173) | (training_df["P2_ID"] == 206173))]
sinner[sinner['TOURNEY_NAME'] == 'US Open'][-10:]

training_df.to_csv("../data/processed/Training_Data.csv", index=False)
